Adicionado metódos na regra de négocio e excluído a classe Status

+++ src/dao/DAOAluno.java
@@ -2,7 +2,6 @@ package dao;
 
 import javax.persistence.PersistenceException;
 import javax.persistence.Query;
-
 import exception.DAOException;
 import model.Aluno;
 
@@ -14,38 +13,43 @@ public class DAOAluno extends DAOGenerico<Aluno> implements IDAOAluno {
 	 * @param turma_id
 	 * @return
 	 */
-	public long QtdAlunoTurma(int turma_id) throws DAOException {
-		long qtd;
+	public int QtdAlunoTurma(int turma_id) throws DAOException {
+		int qtd;
 		try {
 			Query query = em.createQuery("SELECT COUNT (*) from Aluno a WHERE a.id_turma = :id", Aluno.class);
-			query.setParameter("id",turma_id);
-			qtd = (long) query.getSingleResult();
+			query.setParameter("id", turma_id);
+			qtd = (int) query.getSingleResult();
 		} catch (PersistenceException e) {
 			throw new DAOException("Erro ao obter a quantidade de Alunos!");
 		}
 		return qtd;
 	}
-	public long NumeroFaltas(int id_aluno) throws DAOException{
-		long qtd;
-		try{
-			Query query = em.createQuery("SELECT a.faltas FROM Aluno a WHERE a.id = :id",Aluno.class);
-			query.setParameter("id",id_aluno);
-			qtd = (long) query.getSingleResult();
-		}catch(PersistenceException e){
+	/**
+	 * Retorna a quantidade de faltas de um aluno
+	 */
+	public int NumeroFaltas(int id_aluno) throws DAOException {
+		int qtd;
+		try {
+			Query query = em.createQuery("SELECT a.faltas FROM Aluno a WHERE a.id = :id", Aluno.class);
+			query.setParameter("id", id_aluno);
+			qtd = (int) query.getSingleResult();
+		} catch (PersistenceException e) {
 			throw new DAOException("Erro ao verificar a quantidade de faltas!");
 		}
 		return qtd;
 	}
-	public void LancaFalta(Aluno a) throws DAOException{
+
+	public void LancaFalta(Aluno a) throws DAOException {
 		int faltas;
-		try{
+		try {
 			et.begin();
 			faltas = a.getFaltas();
-			a.setFaltas(faltas+1);
+			a.setFaltas(faltas + 1);
 			em.merge(a);
 			et.commit();
-		}catch(PersistenceException e){
+		} catch (PersistenceException e) {
 			throw new DAOException("Erro ao lançar falta");
 		}
 	}
+
 }
