+++ src/dao/IDAOAluno.java
@@ -5,8 +5,8 @@ import model.Aluno;
 
 public interface IDAOAluno extends IDAOGenerico<Aluno>{
 
-	public long QtdAlunoTurma(int turma_id) throws DAOException;
-	public long NumeroFaltas(int id_aluno) throws DAOException;
+	public int QtdAlunoTurma(int turma_id) throws DAOException;
+	public int NumeroFaltas(int id_aluno) throws DAOException;
 	public void LancaFalta(Aluno a) throws DAOException;
 	
 }
